+++ product-service/getProductByIdHandler.mjs
@@ -1,0 +1,23 @@
+import { getProductById } from "./ProductService.mjs";
+
+export const getProductByIdHandler = async (event) => {
+  const productId = event.pathParameters.productId;
+  const product = await getProductById(productId);
+  if (product) {
+    return {
+      statusCode: 200,
+      headers: {
+        "Access-Control-Allow-Origin": process.env.frontendURL,
+      },
+      body: JSON.stringify(product),
+    };
+  } else {
+    return {
+      statusCode: 404,
+      headers: {
+        "Access-Control-Allow-Origin": process.env.frontendURL,
+      },
+      body: JSON.stringify({ message: "Product not found" }),
+    };
+  }
+};
